# Optimize the implementation of the argsort operator.