<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
Optimize the implementation of argsort operator. The performance is better than Pytorch.  
Given on the input size of [4, 68992], which is slower than Pytorch initially, the performance of optimized argsort kernel can achieve 3.3x speedup, which is ~1.5x the performance of Pytorch. 
Reorganize the code of argsort operator to reduce the code redundancy.  
This optimal codes are called when the number of rows of input tensor is less.  
When the number of input tensor is large, the original implementation is called. (The performance is also better than Pytorch)  

| case | pytorch | paddle (before optimization) | diff | paddle (after optimization) | diff | speedup |
|------|--------|-----|------|------|------|-------|
|x(Variable)-dtype: float32, shape: [1L, 68892L]  axis(int):-1 | 0.17 | 0.23 | lower than(35.3%) | 0.20| lower than(17.6%)| 1.15|
|x(Variable)-dtype: float32, shape: [2L, 68892L]  axis(int):-1|0.25|0.91|lower than(2.64)|0.28|lower than(12%)|3.25|
|x(Variable)-dtype: float32, shape: [4L, 68892L]  axis(int):-1|0.42|0.97|lower than(1.31)|0.29|greater than(31.0%)|3.34|
|x(Variable)-dtype: float32, shape: [8L, 68892L]  axis(int):-1|0.71|1.03|lower than(45.1%)|0.39|greater than(45.1%)|2.64|
|x(Variable)-dtype: float32, shape: [16L, 68892L]  axis(int):-1|1.21|1.07|greater than(11.6%)|1.04|greater than(14.0%)|1.03|
|x(Variable)-dtype: float32, shape: [32L, 68892L]  axis(int):-1|2.23|1.31|greater than(41.3%)|1.11|greater than(50.2%)|1.18|